class-service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
python-multipart
httpx
